Move common.py into post_data.py
